简化 IPagingQueryBuilder, 增加 扩展方法 ,支持复杂聚合
